Relocate `PgSearchScanTaskEstimator` next to the thing it is estimating tasks for.
